M24CSA007_A1/requirements.txt
glfw==2.8.0
imgui==2.0.0
numpy==2.2.3
PyOpenGL==3.1.9
PyOpenGL==3.1.9
